Extract sync loop into src/sync_loop.rs
Move the watch-mode sync loop, per-library sync orchestration, and
LibraryState into a dedicated module. Decompose the monolithic loop
into run_sync, run_cycle, check_changes_database, determine_sync_mode,
and reacquire_session helpers.

main.rs: 1457 -> 704 lines. sync_loop.rs: 884 lines.
Overall: main.rs went from 2888 to 704 lines across commits 7-8.
